+++ memapi/src/memorytracking.rs
@@ -10,6 +10,7 @@ use ahash::RandomState as ARandomState;
 use im::Vector as ImVector;
 use itertools::Itertools;
 use std::borrow::Cow;
+use std::collections::BTreeMap;
 use std::collections::HashMap;
 use std::path::Path;
 
@@ -18,16 +19,16 @@ extern "C" {
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct FunctionId(u32);
+pub struct FunctionId(u64);
 
 impl FunctionId {
-    pub const UNKNOWN: Self = Self(u32::MAX);
+    pub const UNKNOWN: Self = Self(u64::MAX);
 
-    pub fn new(id: u32) -> Self {
+    pub fn new(id: u64) -> Self {
         FunctionId(id)
     }
 
-    pub fn as_u32(&self) -> u32 {
+    pub fn as_u64(&self) -> u64 {
         self.0
     }
 }
@@ -39,13 +40,17 @@ struct FunctionLocation {
     function_name: String,
 }
 
+pub trait FunctionLocations {
+    fn get_function_and_filename(&self, id: FunctionId) -> (&str, &str);
+}
+
 /// Stores FunctionLocations, returns a FunctionId
 #[derive(Clone)]
-pub struct FunctionLocations {
+pub struct VecFunctionLocations {
     functions: Vec<FunctionLocation>,
 }
 
-impl FunctionLocations {
+impl VecFunctionLocations {
     /// Create a new tracker.
     pub fn new() -> Self {
         Self {
@@ -61,9 +66,11 @@ impl FunctionLocations {
         });
         // If we ever have 2 ** 32 or more functions in our program, this will
         // break. Seems unlikely, even with long running workers.
-        FunctionId((self.functions.len() - 1) as u32)
+        FunctionId((self.functions.len() - 1) as u64)
     }
+}
 
+impl FunctionLocations for VecFunctionLocations {
     /// Get the function name and filename.
     fn get_function_and_filename(&self, id: FunctionId) -> (&str, &str) {
         if id == FunctionId::UNKNOWN {
@@ -162,7 +169,7 @@ impl Callstack {
         &self,
         to_be_post_processed: bool,
         prefix_stripper: Option<&PrefixStripper>,
-        functions: &FunctionLocations,
+        functions: &dyn FunctionLocations,
         separator: &'static str,
     ) -> String {
         if self.calls.is_empty() {
@@ -249,7 +256,7 @@ pub struct CallstackInterner {
     callstack_to_id: HashMap<Callstack, u32, ARandomState>,
 }
 
-impl<'a> CallstackInterner {
+impl CallstackInterner {
     pub fn new() -> Self {
         CallstackInterner {
             max_id: 0,
@@ -288,6 +295,14 @@ impl<'a> CallstackInterner {
 const MIB: usize = 1024 * 1024;
 const HIGH_32BIT: u32 = 1 << 31;
 
+/// A unique identifier for a process. The idea is that each subprocess will be
+/// given a unique identifier from a counter, and that >4 billion processes is
+/// unlikely. But the internal representation can change!
+#[derive(Clone, Copy, Debug, PartialEq, Ord, PartialOrd, Eq)]
+pub struct ProcessUid(u32);
+
+pub const PARENT_PROCESS: ProcessUid = ProcessUid(0);
+
 /// A specific call to malloc()/calloc().
 #[derive(Clone, Copy, Debug, PartialEq)]
 struct Allocation {
@@ -324,15 +339,15 @@ impl Allocation {
 }
 
 /// The main data structure tracking everything.
-pub struct AllocationTracker {
+pub struct AllocationTracker<FL: FunctionLocations> {
     // malloc()/calloc():
-    current_allocations: HashMap<usize, Allocation, ARandomState>,
+    current_allocations: BTreeMap<ProcessUid, HashMap<usize, Allocation, ARandomState>>,
     // anonymous mmap(), i.e. not file backed:
-    current_anon_mmaps: RangeMap<CallstackId>,
+    current_anon_mmaps: BTreeMap<ProcessUid, RangeMap<CallstackId>>,
 
     // Map FunctionIds to function + filename strings, so we can store the
     // former and save memory.
-    pub functions: FunctionLocations,
+    pub functions: FL,
 
     // Map CallstackIds to Callstacks, so we can store the former and save
     // memory:
@@ -353,15 +368,15 @@ pub struct AllocationTracker {
     failed_deallocations: usize,
 }
 
-impl<'a> AllocationTracker {
-    pub fn new(default_path: String) -> AllocationTracker {
+impl<FL: FunctionLocations> AllocationTracker<FL> {
+    pub fn new(default_path: String, functions: FL) -> AllocationTracker<FL> {
         AllocationTracker {
-            current_allocations: new_hashmap(),
-            current_anon_mmaps: RangeMap::new(),
+            current_allocations: BTreeMap::from([(PARENT_PROCESS, new_hashmap())]),
+            current_anon_mmaps: BTreeMap::from([(PARENT_PROCESS, RangeMap::new())]),
             interner: CallstackInterner::new(),
-            functions: FunctionLocations::new(),
             current_memory_usage: ImVector::new(),
             peak_memory_usage: ImVector::new(),
+            functions,
             current_allocated_bytes: 0,
             peak_allocated_bytes: 0,
             missing_allocated_bytes: 0,
@@ -385,8 +400,13 @@ impl<'a> AllocationTracker {
         self.current_allocated_bytes
     }
 
-    pub fn get_allocation_size(&self, address: usize) -> usize {
-        if let Some(allocation) = self.current_allocations.get(&address) {
+    pub fn get_allocation_size(&self, process: ProcessUid, address: usize) -> usize {
+        if let Some(allocation) = self
+            .current_allocations
+            .get(&process)
+            .map(|a| a.get(&address))
+            .flatten()
+        {
             allocation.size()
         } else {
             0
@@ -424,10 +444,21 @@ impl<'a> AllocationTracker {
     }
 
     /// Add a new allocation based off the current callstack.
-    pub fn add_allocation(&mut self, address: usize, size: usize, callstack_id: CallstackId) {
+    pub fn add_allocation(
+        &mut self,
+        process: ProcessUid,
+        address: usize,
+        size: usize,
+        callstack_id: CallstackId,
+    ) {
         let alloc = Allocation::new(callstack_id, size);
         let compressed_size = alloc.size();
-        if let Some(previous) = self.current_allocations.insert(address, alloc) {
+        if let Some(previous) = self
+            .current_allocations
+            .entry(process)
+            .or_default()
+            .insert(address, alloc)
+        {
             // In production use (proposed commercial product) allocations are
             // only sampled, so missing allocations are common and not the sign
             // of an error.
@@ -461,11 +492,16 @@ impl<'a> AllocationTracker {
     }
 
     /// Free an existing allocation, return how much was removed, if any.
-    pub fn free_allocation(&mut self, address: usize) -> Option<usize> {
+    pub fn free_allocation(&mut self, process: ProcessUid, address: usize) -> Option<usize> {
         // Before we reduce memory, let's check if we've previously hit a peak:
         self.check_if_new_peak();
 
-        if let Some(removed) = self.current_allocations.remove(&address) {
+        if let Some(removed) = self
+            .current_allocations
+            .entry(process)
+            .or_default()
+            .remove(&address)
+        {
             self.remove_memory_usage(removed.callstack_id, removed.size());
             Some(removed.size())
         } else {
@@ -485,17 +521,51 @@ impl<'a> AllocationTracker {
     }
 
     /// Add a new anonymous mmap() based of the current callstack.
-    pub fn add_anon_mmap(&mut self, address: usize, size: usize, callstack_id: CallstackId) {
-        self.current_anon_mmaps.add(address, size, callstack_id);
+    pub fn add_anon_mmap(
+        &mut self,
+        process: ProcessUid,
+        address: usize,
+        size: usize,
+        callstack_id: CallstackId,
+    ) {
+        self.current_anon_mmaps
+            .entry(process)
+            .or_default()
+            .add(address, size, callstack_id);
         self.add_memory_usage(callstack_id, size);
     }
 
-    pub fn free_anon_mmap(&mut self, address: usize, size: usize) {
+    pub fn free_anon_mmap(&mut self, process: ProcessUid, address: usize, size: usize) {
         // Before we reduce memory, let's check if we've previously hit a peak:
         self.check_if_new_peak();
         // Now remove, and update totoal memory tracking:
-        for (callstack_id, removed) in self.current_anon_mmaps.remove(address, size) {
+        for (callstack_id, removed) in self
+            .current_anon_mmaps
+            .entry(process)
+            .or_default()
+            .remove(address, size)
+        {
             self.remove_memory_usage(callstack_id, removed);
+        }
+    }
+
+    /// The process just died, remove all the allocations.
+    pub fn drop_process(&mut self, process: ProcessUid) {
+        // Before we reduce memory, let's check if we've previously hit a peak:
+        self.check_if_new_peak();
+
+        // Drop anon mmaps, call remove_memory_usage on all entries.
+        if let Some(mmaps_for_process) = self.current_anon_mmaps.remove(&process) {
+            for (size, callstack_id) in mmaps_for_process.into_iter() {
+                self.remove_memory_usage(callstack_id, size);
+            }
+        }
+
+        // Drop allocations, call remove_memory_usage on all entries.
+        if let Some(allocations_for_process) = self.current_allocations.remove(&process) {
+            for allocation in allocations_for_process.values() {
+                self.remove_memory_usage(allocation.callstack_id, allocation.size());
+            }
         }
     }
 
@@ -554,21 +624,19 @@ impl<'a> AllocationTracker {
     ) -> impl Iterator<Item = String> + '_ {
         let by_call = self.combine_callstacks(peak).into_iter();
         let id_to_callstack = self.interner.get_reverse_map();
-        let functions = &self.functions;
         let prefix_stripper = PrefixStripper::new();
-        let lines = by_call.map(move |(callstack_id, size)| {
+        by_call.map(move |(callstack_id, size)| {
             format!(
                 "{} {}",
                 id_to_callstack.get(&callstack_id).unwrap().as_string(
                     to_be_post_processed,
                     Some(&prefix_stripper),
-                    functions,
+                    &self.functions,
                     ";"
                 ),
                 size,
             )
-        });
-        lines
+        })
     }
 
     fn dump_to_flamegraph(
@@ -624,7 +692,6 @@ impl<'a> AllocationTracker {
     /// Clear memory we won't be needing anymore, since we're going to exit out.
     pub fn oom_break_glass(&mut self) {
         self.current_allocations.clear();
-        self.current_allocations.shrink_to_fit();
         self.peak_memory_usage.clear();
     }
 
@@ -650,10 +717,15 @@ impl<'a> AllocationTracker {
     /// check_if_new_peak() is called.
     fn validate(&self) {
         assert!(self.peak_allocated_bytes >= self.current_allocated_bytes);
-        let current_allocations = self.current_anon_mmaps.size()
+        let current_allocations: usize = self
+            .current_anon_mmaps
+            .values()
+            .map(|maps| maps.size())
+            .sum::<usize>()
             + self
                 .current_allocations
-                .iter()
+                .values()
+                .flat_map(|allocs| allocs.iter())
                 .map(|(_, alloc)| alloc.size())
                 .sum::<usize>();
         assert!(
@@ -670,7 +742,7 @@ impl<'a> AllocationTracker {
     /// caching of callstack ID.
     pub fn reset(&mut self, default_path: String) {
         self.current_allocations.clear();
-        self.current_anon_mmaps = RangeMap::new();
+        self.current_anon_mmaps = BTreeMap::from([(PARENT_PROCESS, RangeMap::new())]);
         for i in self.current_memory_usage.iter_mut() {
             *i = 0;
         }
@@ -684,13 +756,19 @@ impl<'a> AllocationTracker {
 
 #[cfg(test)]
 mod tests {
+    use crate::memorytracking::{ProcessUid, PARENT_PROCESS};
+
     use super::{
         Allocation, AllocationTracker, CallSiteId, Callstack, CallstackInterner, FunctionId,
-        FunctionLocations, HIGH_32BIT, MIB,
+        FunctionLocations, VecFunctionLocations, HIGH_32BIT, MIB,
     };
     use proptest::prelude::*;
     use std::borrow::Cow;
     use std::collections::HashMap;
+
+    fn new_tracker() -> AllocationTracker<VecFunctionLocations> {
+        AllocationTracker::new(".".to_string(), VecFunctionLocations::new())
+    }
 
     proptest! {
         // Allocation sizes smaller than 2 ** 31 are round-tripped.
@@ -702,7 +780,6 @@ mod tests {
 
         // Allocation sizes larger than 2 ** 31 are stored as MiBs, with some
         // loss of resolution.
-
         #[test]
         fn large_allocation(size in (HIGH_32BIT as usize)..(1 << 50)) {
             let allocation = Allocation::new(0, size as usize);
@@ -718,17 +795,17 @@ mod tests {
         // Test for https://github.com/pythonspeed/filprofiler/issues/66
         #[test]
         fn correct_allocation_size_tracked(size in (1 as usize)..(1<< 50)) {
-            let mut tracker = AllocationTracker::new(".".to_string());
+            let mut tracker = new_tracker();
             let cs_id = tracker.get_callstack_id(&Callstack::new());
-            tracker.add_allocation(0, size, cs_id);
-            tracker.add_anon_mmap(1, size * 2, cs_id);
+            tracker.add_allocation(PARENT_PROCESS, 0, size, cs_id);
+            tracker.add_anon_mmap(PARENT_PROCESS, 1, size * 2, cs_id);
             // We don't track (large) allocations exactly right, but they should
             // be quite close:
             let ratio = ((size * 3) as f64) / (tracker.current_memory_usage[0] as f64);
             prop_assert!(0.999 < ratio);
             prop_assert!(ratio < 1.001);
-            tracker.free_allocation(0);
-            tracker.free_anon_mmap(1, size * 2);
+            tracker.free_allocation(PARENT_PROCESS, 0);
+            tracker.free_anon_mmap(PARENT_PROCESS, 1, size * 2);
             // Once we've freed everything, it should be _exactly_ 0.
             prop_assert_eq!(&im::vector![0], &tracker.current_memory_usage);
             tracker.check_if_new_peak();
@@ -738,29 +815,32 @@ mod tests {
         #[test]
         fn current_allocated_matches_sum_of_allocations(
             // Allocated bytes. Will use index as the memory address.
-            allocated_sizes in prop::collection::vec(1..1000 as usize, 10..20),
+            allocated_sizes in prop::collection::vec((0..2 as u32, 1..100 as usize), 10..20),
             // Allocations to free.
             free_indices in prop::collection::btree_set(0..10 as usize, 1..5)
         ) {
-            let mut tracker = AllocationTracker::new(".".to_string());
+            let mut tracker = new_tracker();
             let mut expected_memory_usage = im::vector![];
             for i in 0..allocated_sizes.len() {
+                let (process, allocation_size) = *allocated_sizes.get(i).unwrap();
+                let process = ProcessUid(process);
                 let mut cs = Callstack::new();
-                cs.start_call(0, CallSiteId::new(FunctionId::new(i as u32), 0));
+                cs.start_call(0, CallSiteId::new(FunctionId::new(i as u64), 0));
                 let cs_id = tracker.get_callstack_id(&cs);
-                tracker.add_allocation(i as usize,*allocated_sizes.get(i).unwrap(), cs_id);
-                expected_memory_usage.push_back(*allocated_sizes.get(i).unwrap());
+                tracker.add_allocation(process, i as usize, allocation_size, cs_id);
+                expected_memory_usage.push_back(allocation_size);
             }
-            let mut expected_sum = allocated_sizes.iter().sum();
+            let mut expected_sum = allocated_sizes.iter().map(|t| t.1).sum();
             let expected_peak : usize = expected_sum;
             prop_assert_eq!(tracker.current_allocated_bytes, expected_sum);
             prop_assert_eq!(&tracker.current_memory_usage, &expected_memory_usage);
             for i in free_indices.iter() {
-                let expected_removed = allocated_sizes.get(*i).unwrap();
+                let (process, expected_removed) = allocated_sizes.get(*i).unwrap();
+                let process = ProcessUid(*process);
                 expected_sum -= expected_removed;
-                let removed = tracker.free_allocation(*i);
+                let removed = tracker.free_allocation(process, *i);
                 prop_assert_eq!(removed, Some(*expected_removed));
-                expected_memory_usage[*i] -= allocated_sizes.get(*i).unwrap();
+                expected_memory_usage[*i] -= expected_removed;
                 prop_assert_eq!(tracker.current_allocated_bytes, expected_sum);
                 prop_assert_eq!(&tracker.current_memory_usage, &expected_memory_usage);
             }
@@ -772,29 +852,33 @@ mod tests {
         #[test]
         fn current_allocated_anon_maps_matches_sum_of_allocations(
             // Allocated bytes. Will use index as the memory address.
-            allocated_sizes in prop::collection::vec(1..1000 as usize, 10..20),
+            allocated_sizes in prop::collection::vec((0..2 as u32, 1..100 as usize), 10..20),
             // Allocations to free.
             free_indices in prop::collection::btree_set(0..10 as usize, 1..5)
         ) {
-            let mut tracker = AllocationTracker::new(".".to_string());
+            let mut tracker = new_tracker();
             let mut expected_memory_usage = im::vector![];
             // Make sure addresses don't overlap:
             let addresses : Vec<usize> = (0..allocated_sizes.len()).map(|i| i * 10000).collect();
             for i in 0..allocated_sizes.len() {
+                let (process, allocation_size) = *allocated_sizes.get(i).unwrap();
+                let process = ProcessUid(process);
                 let mut cs = Callstack::new();
-                cs.start_call(0, CallSiteId::new(FunctionId::new(i as u32), 0));
+                cs.start_call(0, CallSiteId::new(FunctionId::new(i as u64), 0));
                 let csid = tracker.get_callstack_id(&cs);
-                tracker.add_anon_mmap(addresses[i] as usize, *allocated_sizes.get(i).unwrap(), csid);
-                expected_memory_usage.push_back(*allocated_sizes.get(i).unwrap());
+                tracker.add_anon_mmap(process, addresses[i] as usize, allocation_size, csid);
+                expected_memory_usage.push_back(allocation_size);
             }
-            let mut expected_sum = allocated_sizes.iter().sum();
+            let mut expected_sum = allocated_sizes.iter().map(|t|t.1).sum();
             let expected_peak : usize = expected_sum;
             prop_assert_eq!(tracker.current_allocated_bytes, expected_sum);
             prop_assert_eq!(&tracker.current_memory_usage, &expected_memory_usage);
             for i in free_indices.iter() {
-                expected_sum -= allocated_sizes.get(*i).unwrap();
-                tracker.free_anon_mmap(addresses[*i], *allocated_sizes.get(*i).unwrap());
-                expected_memory_usage[*i] -= allocated_sizes.get(*i).unwrap();
+                let (process, allocation_size) = *allocated_sizes.get(*i).unwrap();
+                let process = ProcessUid(process);
+                expected_sum -= allocation_size;
+                tracker.free_anon_mmap(process, addresses[*i], allocation_size);
+                expected_memory_usage[*i] -= allocation_size;
                 prop_assert_eq!(tracker.current_allocated_bytes, expected_sum);
                 prop_assert_eq!(&tracker.current_memory_usage, &expected_memory_usage);
             }
@@ -802,19 +886,60 @@ mod tests {
             tracker.check_if_new_peak();
             tracker.validate();
         }
+
+        #[test]
+        fn drop_process_removes_that_process_allocations_and_mmaps(
+            // Allocated bytes. Will use index as the memory address.
+            allocated_sizes in prop::collection::vec((0..2 as u32, 1..100 as usize), 10..20),
+            allocated_mmaps in prop::collection::vec((0..2 as u32, 1..100 as usize), 10..20),
+        ) {
+            let mut tracker = new_tracker();
+            let mut expected_memory_usage : usize = 0;
+            // Make sure addresses don't overlap:
+            let mmap_addresses : Vec<usize> = (0..allocated_mmaps.len()).map(|i| i * 10000).collect();
+            for i in 0..allocated_sizes.len() {
+                let (process, allocation_size) = *allocated_sizes.get(i).unwrap();
+                let process = ProcessUid(process);
+                let mut cs = Callstack::new();
+                cs.start_call(0, CallSiteId::new(FunctionId::new(i as u64), 0));
+                let cs_id = tracker.get_callstack_id(&cs);
+                tracker.add_allocation(process, i as usize, allocation_size, cs_id);
+                expected_memory_usage += allocation_size;
+            }
+            for i in 0..allocated_mmaps.len() {
+                let (process, allocation_size) = *allocated_mmaps.get(i).unwrap();
+                let process = ProcessUid(process);
+                let mut cs = Callstack::new();
+                cs.start_call(0, CallSiteId::new(FunctionId::new(i as u64), 0));
+                let csid = tracker.get_callstack_id(&cs);
+                tracker.add_anon_mmap(process, mmap_addresses[i] as usize, allocation_size, csid);
+                expected_memory_usage += allocation_size;
+            }
+            prop_assert_eq!(tracker.current_allocated_bytes, expected_memory_usage);
+            let expected_peak = expected_memory_usage;
+            let to_drop = ProcessUid(1);
+            tracker.drop_process(to_drop);
+            expected_memory_usage -= allocated_sizes.iter().filter(|(p, _)| ProcessUid(*p) == to_drop).map(|(_, size)| size).sum::<usize>();
+            expected_memory_usage -= allocated_mmaps.iter().filter(|(p, _)| ProcessUid(*p) == to_drop).map(|(_, size)| size).sum::<usize>();
+            prop_assert_eq!(tracker.current_allocated_bytes, expected_memory_usage);
+            prop_assert_eq!(tracker.peak_allocated_bytes, expected_peak);
+            tracker.check_if_new_peak();
+            tracker.validate();
+        }
+
     }
 
     #[test]
     fn untracked_allocation_removal() {
-        let mut tracker = AllocationTracker::new("/tmp".to_string());
-        assert_eq!(tracker.free_allocation(123), None);
+        let mut tracker = new_tracker();
+        assert_eq!(tracker.free_allocation(PARENT_PROCESS, 123), None);
     }
 
     #[test]
     fn callstack_line_numbers() {
-        let fid1 = FunctionId::new(1u32);
-        let fid3 = FunctionId::new(3u32);
-        let fid5 = FunctionId::new(5u32);
+        let fid1 = FunctionId::new(1u64);
+        let fid3 = FunctionId::new(3u64);
+        let fid5 = FunctionId::new(5u64);
 
         // Parent line number does nothing if it's first call:
         let mut cs1 = Callstack::new();
@@ -841,8 +966,8 @@ mod tests {
 
     #[test]
     fn callstackinterner_notices_duplicates() {
-        let fid1 = FunctionId::new(1u32);
-        let fid3 = FunctionId::new(3u32);
+        let fid1 = FunctionId::new(1u64);
+        let fid3 = FunctionId::new(3u64);
 
         let mut cs1 = Callstack::new();
         cs1.start_call(0, CallSiteId::new(fid1, 2));
@@ -897,7 +1022,7 @@ mod tests {
             cs1.id_for_new_allocation(0, |cs| interner.get_or_insert_id(Cow::Borrowed(&cs), || ()));
         assert_eq!(id0, id0b);
 
-        let fid1 = FunctionId::new(1u32);
+        let fid1 = FunctionId::new(1u64);
 
         cs1.start_call(0, CallSiteId::new(fid1, 2));
         let id1 =
@@ -943,10 +1068,10 @@ mod tests {
 
     #[test]
     fn peak_allocations_only_updated_on_new_peaks() {
-        let fid1 = FunctionId::new(1u32);
-        let fid3 = FunctionId::new(3u32);
+        let fid1 = FunctionId::new(1u64);
+        let fid3 = FunctionId::new(3u64);
 
-        let mut tracker = AllocationTracker::new(".".to_string());
+        let mut tracker = new_tracker();
         let mut cs1 = Callstack::new();
         cs1.start_call(0, CallSiteId::new(fid1, 2));
         let mut cs2 = Callstack::new();
@@ -954,7 +1079,7 @@ mod tests {
 
         let cs1_id = tracker.get_callstack_id(&cs1);
 
-        tracker.add_allocation(1, 1000, cs1_id);
+        tracker.add_allocation(PARENT_PROCESS, 1, 1000, cs1_id);
         tracker.check_if_new_peak();
         // Peak should now match current allocations:
         assert_eq!(tracker.current_memory_usage, im::vector![1000]);
@@ -963,14 +1088,14 @@ mod tests {
         let previous_peak = tracker.peak_memory_usage.clone();
 
         // Free the allocation:
-        tracker.free_allocation(1);
+        tracker.free_allocation(PARENT_PROCESS, 1);
         assert_eq!(tracker.current_allocated_bytes, 0);
         assert_eq!(tracker.current_memory_usage, im::vector![0]);
         assert_eq!(previous_peak, tracker.peak_memory_usage);
         assert_eq!(tracker.peak_allocated_bytes, 1000);
 
         // Add allocation, still less than 1000:
-        tracker.add_allocation(3, 123, cs1_id);
+        tracker.add_allocation(PARENT_PROCESS, 3, 123, cs1_id);
         assert_eq!(tracker.current_memory_usage, im::vector![123]);
         tracker.check_if_new_peak();
         assert_eq!(previous_peak, tracker.peak_memory_usage);
@@ -978,7 +1103,7 @@ mod tests {
 
         // Add allocation that goes past previous peak
         let cs2_id = tracker.get_callstack_id(&cs2);
-        tracker.add_allocation(2, 2000, cs2_id);
+        tracker.add_allocation(PARENT_PROCESS, 2, 2000, cs2_id);
         tracker.check_if_new_peak();
         assert_eq!(tracker.current_memory_usage, im::vector![123, 2000]);
         assert_eq!(tracker.current_memory_usage, tracker.peak_memory_usage);
@@ -986,20 +1111,20 @@ mod tests {
         let previous_peak = tracker.peak_memory_usage.clone();
 
         // Add anonymous mmap() that doesn't go past previous peak:
-        tracker.free_allocation(2);
+        tracker.free_allocation(PARENT_PROCESS, 2);
         assert_eq!(tracker.current_memory_usage, im::vector![123, 0]);
-        tracker.add_anon_mmap(50000, 1000, cs2_id);
+        tracker.add_anon_mmap(PARENT_PROCESS, 50000, 1000, cs2_id);
         assert_eq!(tracker.current_memory_usage, im::vector![123, 1000]);
         tracker.check_if_new_peak();
         assert_eq!(tracker.current_allocated_bytes, 1123);
         assert_eq!(tracker.peak_allocated_bytes, 2123);
         assert_eq!(tracker.peak_memory_usage, previous_peak);
         assert_eq!(tracker.current_allocations.len(), 1);
-        assert!(tracker.current_allocations.contains_key(&3));
-        assert!(tracker.current_anon_mmaps.size() > 0);
+        assert!(tracker.current_allocations[&PARENT_PROCESS].contains_key(&3));
+        assert!(tracker.current_anon_mmaps[&PARENT_PROCESS].size() > 0);
 
         // Add anonymous mmap() that does go past previous peak:
-        tracker.add_anon_mmap(600000, 2000, cs2_id);
+        tracker.add_anon_mmap(PARENT_PROCESS, 600000, 2000, cs2_id);
         assert_eq!(tracker.current_memory_usage, im::vector![123, 3000]);
         tracker.check_if_new_peak();
         assert_eq!(tracker.current_memory_usage, tracker.peak_memory_usage);
@@ -1007,23 +1132,22 @@ mod tests {
         assert_eq!(tracker.peak_allocated_bytes, 3123);
 
         // Remove mmap():
-        tracker.free_anon_mmap(50000, 1000);
+        tracker.free_anon_mmap(PARENT_PROCESS, 50000, 1000);
         assert_eq!(tracker.current_memory_usage, im::vector![123, 2000]);
         tracker.check_if_new_peak();
         assert_eq!(tracker.current_allocated_bytes, 2123);
         assert_eq!(tracker.peak_allocated_bytes, 3123);
-        assert_eq!(tracker.current_anon_mmaps.size(), 2000);
-        assert!(tracker
-            .current_anon_mmaps
+        assert_eq!(tracker.current_anon_mmaps[&PARENT_PROCESS].size(), 2000);
+        assert!(tracker.current_anon_mmaps[&PARENT_PROCESS]
             .as_hashmap()
             .contains_key(&600000));
 
         // Partial removal of anonmyous mmap():
-        tracker.free_anon_mmap(600100, 1000);
+        tracker.free_anon_mmap(PARENT_PROCESS, 600100, 1000);
         assert_eq!(tracker.current_memory_usage, im::vector![123, 1000]);
         assert_eq!(tracker.current_allocated_bytes, 1123);
         assert_eq!(tracker.peak_allocated_bytes, 3123);
-        assert_eq!(tracker.current_anon_mmaps.size(), 1000);
+        assert_eq!(tracker.current_anon_mmaps[&PARENT_PROCESS].size(), 1000);
         tracker.check_if_new_peak();
         tracker.validate();
     }
@@ -1031,7 +1155,7 @@ mod tests {
     #[test]
     fn combine_callstacks_and_sum_allocations() {
         pyo3::prepare_freethreaded_python();
-        let mut tracker = AllocationTracker::new(".".to_string());
+        let mut tracker = new_tracker();
         let fid1 = tracker
             .functions
             .add_function("a".to_string(), "af".to_string());
@@ -1059,10 +1183,10 @@ mod tests {
         let cs1_id = tracker.get_callstack_id(&cs1);
         let cs2_id = tracker.get_callstack_id(&cs2);
         let cs3_id = tracker.get_callstack_id(&cs3);
-        tracker.add_allocation(1, 1000, cs1_id);
-        tracker.add_allocation(2, 234, cs2_id);
-        tracker.add_anon_mmap(3, 50000, cs1_id);
-        tracker.add_allocation(4, 6000, cs3_id);
+        tracker.add_allocation(PARENT_PROCESS, 1, 1000, cs1_id);
+        tracker.add_allocation(PARENT_PROCESS, 2, 234, cs2_id);
+        tracker.add_anon_mmap(PARENT_PROCESS, 3, 50000, cs1_id);
+        tracker.add_allocation(PARENT_PROCESS, 4, 6000, cs3_id);
 
         // Make sure we notice new peak.
         tracker.check_if_new_peak();
@@ -1094,7 +1218,7 @@ mod tests {
 
     #[test]
     fn test_unknown_function_id() {
-        let func_locations = FunctionLocations::new();
+        let func_locations = VecFunctionLocations::new();
         let (function, filename) = func_locations.get_function_and_filename(FunctionId::UNKNOWN);
         assert_eq!(filename, "UNKNOWN DUE TO BUG");
         assert_eq!(function, "UNKNOWN");